requirements.txt
gradio
httpx
python-dotenv
toml
pydantic
tenacity
